model/insetGAN/requirements.txt
charset-normalizer==2.0.12
cycler==0.11.0
facenet-pytorch==2.5.2
fonttools==4.33.3
idna==3.3
kiwisolver==1.4.2
lpips==0.1.4
matplotlib==3.5.2
numpy==1.21.6
packaging==21.3
pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.0
scipy==1.7.3
six==1.16.0
tqdm==4.64.0
typing-extensions==4.2.0
urllib3==1.26.9
